+++ internal/mode/usage/markdown/spec_duration_test.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/dogmatiq/ferrite"
 	. "github.com/dogmatiq/ferrite/internal/mode/usage/markdown"
-	"github.com/dogmatiq/ferrite/internal/variable"
 	. "github.com/onsi/ginkgo/v2"
 )
 
@@ -20,7 +19,7 @@ var _ = DescribeTable(
 	Entry(
 		"deprecated",
 		"deprecated.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				Deprecated(ferrite.WithRegistry(reg))
@@ -30,7 +29,7 @@ var _ = DescribeTable(
 	Entry(
 		"optional",
 		"optional.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				Optional(ferrite.WithRegistry(reg))
@@ -39,7 +38,7 @@ var _ = DescribeTable(
 	Entry(
 		"required",
 		"required.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				Required(ferrite.WithRegistry(reg))
@@ -48,7 +47,7 @@ var _ = DescribeTable(
 	Entry(
 		"optional with default value",
 		"with-default.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				WithDefault(10 * time.Millisecond).
@@ -58,7 +57,7 @@ var _ = DescribeTable(
 	Entry(
 		"required with default value",
 		"with-default.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				WithDefault(10 * time.Millisecond).
@@ -68,7 +67,7 @@ var _ = DescribeTable(
 	Entry(
 		"with maximum value",
 		"with-max.md",
-		func(reg *variable.Registry) {
+		func(reg ferrite.Registry) {
 			ferrite.
 				Duration("GRPC_TIMEOUT", "gRPC request timeout").
 				WithMaximum(24 * time.Hour).