+++ option_withregistry.go
@@ -2,8 +2,8 @@ package ferrite
 
 import "github.com/dogmatiq/ferrite/internal/variable"
 
-// WithRegistry is an option that sets the variable registry to use.
-func WithRegistry(reg *variable.Registry) interface {
+// WithRegistry is an option that sets the variable registries to use.
+func WithRegistry(reg Registry) interface {
 	InitOption
 	RequiredOption
 	OptionalOption
@@ -13,12 +13,14 @@ func WithRegistry(reg *variable.Registry) interface {
 		panic("registry must not be nil")
 	}
 
+	exposed := variable.ExposeRegistry(reg)
+
 	return option{
 		ApplyToInitConfig: func(cfg *initConfig) {
-			cfg.ModeConfig.Registry = reg
+			cfg.ModeConfig.Registries.Add(exposed)
 		},
 		ApplyToSetConfig: func(cfg *variableSetConfig) {
-			cfg.Registry = reg
+			cfg.Registries = append(cfg.Registries, exposed)
 		},
 	}
 }